==> 04 model interpretation.Rmd <==
---
title: "model interpretations"
author: "Brenna Kelly"
date: "2023-03-08"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(INLA)
```

# model results

```{r}
fixed <- data.frame(res$summary.fixed)

fixed <- tibble::rownames_to_column(fixed, "coef")

fixed_z <- fixed %>%
  filter(grepl("z", coef))
fixed_y <- fixed %>%
  filter(grepl("y", coef))
```


# Binomial
## probability for effects increase 

```{r}
fx_list = vector("list", length = 10)
for(i in c(2:10)) {
  int <- fixed_z[which(fixed_z$coef == "mu_z"), "mean"]
  fx <- fixed_z[i, "mean"]
  coef <- fixed_z[i, 1]
  prop_fx <- exp(int + fx) / (1 + exp(int + fx))
  ci_025 <- exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.025quant"] + fixed_z[i, "X0.025quant"]) / 
    (1 + exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.025quant"] + fixed_z[i, "X0.025quant"]))
  ci_975 <- exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.975quant"] + fixed_z[i, "X0.975quant"]) /
    (1 + exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.975quant"] + fixed_z[i, "X0.975quant"]))
  sig <- ifelse((fixed_z[i, "X0.025quant"] > 0 & fixed_z[i, "X0.975quant"] > 0) |
                  (fixed_z[i, "X0.025quant"] < 0 & fixed_z[i, "X0.975quant"] < 0),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_z[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = coef, 
                   prob = prop_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  fx_list[[i]] <- df
}
fx_list = do.call(rbind, fx_list)
fx_list

data.frame(fixed_z[, 1], 
           prob = round(exp(fixed_z[, 2:8] ) / (1 + exp(fixed_z[, 2:8])), 3))

round(exp(fixed_z$X0.975quant), 2)

# NOT log transformed
# Gini
exp(fixed_z[1, 2:8]) / (1 + exp(fixed_z[1, 2:8]))
exp(fixed_z[3, 2:8] + fixed_z[1, 2:8]) / (1 + exp(fixed_z[3, 2:8] + fixed_z[1, 2:8]))
exp(fixed_z[3, 2:8]) / (1 + exp(fixed_z[3, 2:8]))
exp(fixed_z[4, 2:8] + fixed_z[1, 2:8]) / (1 + exp(fixed_z[4, 2:8] + fixed_z[1, 2:8]))

for(i in c(1:10)) {
  mean <- logit2prob(fixed_z[i, "mean"])
  lower <- logit2prob(fixed_z[i, "X0.025quant"])
  upper <- logit2prob(fixed_z[i, "X0.975quant"])
  df <- data.frame(i = fixed_z[i, 1],
                   mean = mean,
                   lower = lower,
                   upper = upper)
  print(df)
}




```

## probability for increase in simple bin interactions

```{r}
intx_list = vector("list", length = 17)
for(i in c(11:17)) {
  int <- fixed_z[which(fixed_z$coef == "mu_z"), "mean"]
  intx_name_1 <- substr(fixed_z[i, "coef"], 1, 7)
  intx_name_2 <- substr(fixed_z[i, "coef"], 9, 15)
  #print(intx_name_2)
  intx <- fixed_z[i, "mean"]
  intx_name <- ifelse(intx_name_1 == "x_pov_z", intx_name_2, intx_name_1)
  fx_1 <- fixed_z[which(fixed_z$coef == intx_name), "mean"]
  fx_2 <- fixed_z[which(fixed_z$coef == "x_pov_z"), "mean"]
  
  prop_fx <- exp(int + intx + fx_1 + fx_2) / (1 + exp(int + intx + fx_1 + fx_2))
  ci_025 <- exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.025quant"] +
                  fixed_z[i, "X0.025quant"] +
                  fixed_z[which(fixed_z$coef == intx_name), "X0.025quant"] +
                  fixed_z[which(fixed_z$coef == "x_pov_z"), "X0.025quant"]) /
    (1 + exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.025quant"] +
                  fixed_z[i, "X0.025quant"] +
                  fixed_z[which(fixed_z$coef == intx_name), "X0.025quant"] +
                  fixed_z[which(fixed_z$coef == "x_pov_z"), "X0.025quant"]))
  ci_975 <- exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.975quant"] +
                  fixed_z[i, "X0.975quant"] +
                  fixed_z[which(fixed_z$coef == intx_name), "X0.975quant"] +
                  fixed_z[which(fixed_z$coef == "x_pov_z"), "X0.975quant"]) /
    (1 + exp(fixed_z[which(fixed_z$coef == "mu_z"), "X0.975quant"] +
                  fixed_z[i, "X0.975quant"] +
                  fixed_z[which(fixed_z$coef == intx_name), "X0.975quant"] +
                  fixed_z[which(fixed_z$coef == "x_pov_z"), "X0.975quant"]))
  sig <- ifelse((fixed_z[i, "X0.025quant"] > 0 & fixed_z[i, "X0.975quant"] > 0) |
                  (fixed_z[i, "X0.025quant"] < 0 & fixed_z[i, "X0.975quant"] < 0),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_z[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = fixed_z[i, "coef"], 
                   prob = prop_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  intx_list[[i]] <- df
}
intx_list = do.call(rbind, intx_list)
intx_list


for(i in c(11:17)) {
  mean <- logit2prob(fixed_z[i, "mean"])
  lower <- logit2prob(fixed_z[i, "X0.025quant"])
  upper <- logit2prob(fixed_z[i, "X0.975quant"])
  df <- data.frame(i = fixed_z[i, 1],
                   mean = mean,
                   lower = lower,
                   upper = upper)
  print(df)
}



```

## binomial odds (not prob)
```{r}
fixed_z

logit2odds <- function(logit){
  odds <- exp(logit)
  #prob <- odds / (1 + odds)
  return(odds)
}

bin_list = vector("list", length = 17)

for(i in c(1:17)) {
  mean <- logit2odds(fixed_z[i, "mean"])
  lower <- logit2odds(fixed_z[i, "X0.025quant"])
  upper <- logit2odds(fixed_z[i, "X0.975quant"])
  df <- data.frame(i = fixed_z[i, 1],
                   mean = mean,
                   lower = lower,
                   upper = upper)
  bin_list[[i]] <- df
}

bin_list = do.call(rbind, bin_list)
bin_list

```

### broader credible intervals

```{r}

exp(-0.041353502 + 0.038138741 + 0.077950874) # mean oth*pov
exp(-0.311989882 + -0.003340407 + 0.001418131) # 0.025 oth*pov
exp(0.22928288 + 0.07961789 + 0.15448362) # 0.975 oth*pov
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_oth_z`)

#AIAN
exp(-0.041353502 + 0.038138741 + 0.077950874) # mean oth*pov
exp(-0.311989882 + -0.003340407 + 0.001418131) # 0.025 oth*pov
exp(0.22928288 + 0.07961789 + 0.15448362) # 0.975 oth*pov

exp(-0.041353502 + 0.108901134 + 0.105493158) # mean tom*pov
exp(-0.311989882 + -0.054563600 + -0.037914950) # 0.025 tom*pov
exp(0.22928288 + 0.27236587 + 0.24890127) # 0.975 tom*pov
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_tom_z`)

exp(-0.041353502 + 0.041127873 + 0.010139907) # mean nhpi*pov
exp(-0.311989882 + 0.011073147 + -0.046888690) # 0.025 nhpi*pov
exp(0.22928288 + 0.07118260 + 0.06716851) # 0.975 nhpi*pov
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_nhp_z`)

exp(-0.041353502 + 0.518413456 + -0.113499377) # mean black*pov
exp(-0.311989882 + 0.401716520 + -0.243593384) # 0.025 black*pov
exp(0.22928288 + 0.63511039 + 0.01659463) # 0.975 black*pov
inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_bla_z`) # (95% CI [0.86, 2.41])

exp(-0.041353502 + 0.210238161 + -0.044008002) # mean asian*pov
exp(-0.311989882 + 0.122456617 + -0.151234734) # 0.025 asian*pov
exp(0.22928288 + 0.29801970 + 0.06321873) # 0.975 asian*pov
inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_asi_z`)

#x_aia_z
exp(fixed_y[which(fixed_z$coef == "x_pov_z"), "mean"] + 
      fixed_y[which(fixed_z$coef == "x_aia_z"), "mean"] + 
      fixed_y[which(fixed_z$coef == "x_pov_z:x_aia_z"), "mean"]) # mean x_aia_z*pov
exp(fixed_y[which(fixed_z$coef == "x_pov_z"), "X0.025quant"] +
      fixed_y[which(fixed_z$coef == "x_aia_z"), "X0.025quant"] + 
      fixed_y[which(fixed_z$coef == "x_pov_z:x_aia_z"), "X0.025quant"]) # 0.025 x_aia_z*pov
exp(fixed_y[which(fixed_z$coef == "x_pov_z"), "X0.975quant"] +
      fixed_y[which(fixed_z$coef == "x_aia_z"), "X0.975quant"] +
    fixed_y[which(fixed_z$coef == "x_pov_z:x_aia_z"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_aia_z`)

1 - inla.pmarginal(0, res$marginals.fixed$x_bla_z)
inla.pmarginal(0.67, res$marginals.fixed$x_bla_z)
inla.pmarginal(0.89, res$marginals.fixed$x_bla_z)
inla.pmarginal(0.99, res$marginals.fixed$x_bla_z)

1 - inla.pmarginal(0, res$marginals.fixed$x_tom_z)
1 - inla.pmarginal(0, res$marginals.fixed$x_oth_z)
1 - inla.pmarginal(0, res$marginals.fixed$x_asi_z)
1 - inla.pmarginal(0, res$marginals.fixed$x_nhp_z)
inla.pmarginal(0, res$marginals.fixed$x_his_z)
inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_his_z`)
inla.pmarginal(0, res$marginals.fixed$x_aia_z)
inla.pmarginal(0, res$marginals.fixed$x_pov_z)
inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_bla_z`)
inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_asi_z`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_oth_z`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_aia_z`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_nhp_z`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_z:x_tom_z`)

```


## combined binomial fixed effects

```{r}
fixed_bin_list <- rbind(fx_list, intx_list)
fixed_bin_list

write.csv(fixed_bin_list, "fixed_bin_list.csv")
```


# Gamma
## probability increase for simple gamma effects

```{r}
fx_list = vector("list", length = 11)

for(i in c(2:10)) {
  int <- fixed_y[which(fixed_y$coef == "mu_y"), "mean"]
  fx <- fixed_y[i, "mean"]
  coef <- fixed_y[i, 1]
  perc_fx <- log(int + fx)
  ci_025 <- log(fixed_y[i, "X0.025quant"] + fixed_y[which(fixed_y$coef == "mu_y"), "X0.025quant"])
  ci_975 <- log(fixed_y[i, "X0.975quant"] + fixed_y[which(fixed_y$coef == "mu_y"), "X0.975quant"])
  sig <- ifelse((fixed_y[i, "X0.025quant"] > 1.5 & fixed_y[i, "X0.975quant"] > 1.5) |
                  (fixed_y[i, "X0.025quant"] < 1.5 & fixed_y[i, "X0.975quant"] < 1.5),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_y[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = coef, 
                   percent = perc_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  fx_list[[i]] <- df
}
#fx_list_2 = do.call(rbind, fx_list)
#fx_list_2
fx_list = do.call(rbind, fx_list)
fx_list


data.frame(fixed_y[, 1], 
           round(exp(fixed_y[, 2:8]), 3))


```



## probability for interactions

```{r}
intx_list = vector("list", length = 17)
for(i in c(11:17)) {
  int <- fixed_y[which(fixed_y$coef == "mu_y"), "mean"]
  intx_name_1 <- substr(fixed_y[i, "coef"], 1, 7)
  intx_name_2 <- substr(fixed_y[i, "coef"], 9, 15)
  #print(intx_name_2)
  intx <- fixed_y[i, "mean"]
  intx_name <- ifelse(intx_name_1 == "x_pov_y", intx_name_2, intx_name_1)
  fx_1 <- fixed_y[which(fixed_y$coef == intx_name), "mean"]
  fx_2 <- fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"]
  
  perc_fx <- log(fixed_y[i, "mean"] + int)# / + fx_1 + fx_2)# / (1 + exp(int + intx + fx_1 + fx_2))
  ci_025 <- log(fixed_y[i, "X0.025quant"] + fixed_y[which(fixed_y$coef == "mu_y"), "X0.025quant"])
  ci_975 <- log(fixed_y[i, "X0.975quant"] + fixed_y[which(fixed_y$coef == "mu_y"), "X0.975quant"])
  sig <- ifelse((fixed_y[i, "X0.025quant"] > 0 & fixed_y[i, "X0.975quant"] > 0) |
                  (fixed_y[i, "X0.025quant"] < 0 & fixed_y[i, "X0.975quant"] < 0),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_y[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = fixed_y[i, "coef"], 
                   percent = perc_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  intx_list[[i]] <- df
}
intx_list = do.call(rbind, intx_list)
intx_list
fixed_y$coef
round(exp(fixed_y$X0.025quant), 2)

round(res$summary.fixed, 3)

```

## gamma odds (not prob)
```{r}
fixed_y

gam_list = vector("list", length = 17)

for(i in c(1:17)) {
  mean <- exp(fixed_y[i, "mean"])
  lower <- exp(fixed_y[i, "X0.025quant"])
  upper <- exp(fixed_y[i, "X0.975quant"])
  df <- data.frame(i = fixed_y[i, 1],
                   mean = round(mean, 2),
                   lower = round(lower, 2),
                   upper = round(upper, 2))
  gam_list[[i]] <- df
}
gam_list = do.call(rbind, gam_list)

gam_list

```

### broader credible intervals

```{r}

#bla
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_bla_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_bla_y"), "mean"]) # mean bla*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_bla_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_bla_y"), "X0.025quant"]) # 0.025 bla*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_bla_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_bla_y"), "X0.975quant"]) # 0.975 bla*pov
inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_bla_y`)

#x_aia_y
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_aia_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_aia_y"), "mean"]) # mean x_aia_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_aia_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_aia_y"), "X0.025quant"]) # 0.025 x_aia_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_aia_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_aia_y"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_aia_y`)

#x_asi_y
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_asi_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_asi_y"), "mean"]) # mean x_asi_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_asi_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_asi_y"), "X0.025quant"]) # 0.025 x_asi_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_asi_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_asi_y"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_asi_y`)

#x_nhp_y
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_nhp_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_nhp_y"), "mean"]) # mean x_nhp_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_nhp_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_nhp_y"), "X0.025quant"]) # 0.025 x_nhp_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_nhp_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_nhp_y"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_nhp_y`)

#x_tom_y
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_tom_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_tom_y"), "mean"]) # mean x_tom_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_tom_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_tom_y"), "X0.025quant"]) # 0.025 x_tom_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_tom_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_tom_y"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_tom_y`)

#x_oth_y
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_oth_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_oth_y"), "mean"]) # mean x_oth_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_oth_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_oth_y"), "X0.025quant"]) # 0.025 x_oth_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_oth_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_oth_y"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_oth_y`)

# x_his_y
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_his_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_his_y"), "mean"]) # mean x_his_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_his_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_his_y"), "X0.025quant"]) # 0.025 x_his_y*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_his_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_his_y"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_his_y`)

exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_bla_y"), "mean"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_bla_y"), "mean"]) # mean bla*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.025quant"] +
      fixed_y[which(fixed_y$coef == "x_bla_y"), "X0.025quant"] + 
      fixed_y[which(fixed_y$coef == "x_pov_y:x_bla_y"), "X0.025quant"]) # 0.025 bla*pov
exp(fixed_y[which(fixed_y$coef == "x_pov_y"), "X0.975quant"] +
      fixed_y[which(fixed_y$coef == "x_bla_y"), "X0.975quant"] +
    fixed_y[which(fixed_y$coef == "x_pov_y:x_bla_y"), "X0.975quant"])
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_oth_y`)

1 - inla.pmarginal(0, res$marginals.fixed$x_pov_y)
1 - inla.pmarginal(0, res$marginals.fixed$x_bla_y)
1 - inla.pmarginal(0, res$marginals.fixed$x_asi_y)
inla.pmarginal(0, res$marginals.fixed$x_nhp_y)
inla.pmarginal(0, res$marginals.fixed$x_aia_y)
inla.pmarginal(0, res$marginals.fixed$x_tom_y)
1 - inla.pmarginal(0, res$marginals.fixed$x_oth_y)
1 - inla.pmarginal(0, res$marginals.fixed$x_his_y)
inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_bla_y`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_aia_y`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_asi_y`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_nhp_y`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_tom_y`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_oth_y`)
1 - inla.pmarginal(0, res$marginals.fixed$`x_pov_y:x_his_y`)




```



## combined gamma fixed effects

```{r}
fixed_gam_list <- rbind(fx_list, intx_list)
fixed_gam_list

write.csv(fixed_gam_list, "fixed_gam_list.csv")
```
