==> 04b cancer model interaction.Rmd <==
---
title: "model interpretations"
author: "Brenna Kelly"
date: "2023-03-08"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# model results

```{r}
fixed_cancer <- data.frame(res_cancer$summary.fixed)

fixed_cancer <- tibble::rownames_to_column(fixed_cancer, "coef")

fixed_cancer_z <- fixed_cancer %>%
  filter(grepl("z", coef))
fixed_cancer_y <- fixed_cancer %>%
  filter(grepl("y", coef))
```


# Binomial
## probability for effects increase 

```{r}
fx_list = vector("list", length = 10)
for(i in c(2:10)) {
  int <- fixed_cancer_z[which(fixed_z$coef == "mu_z"), "mean"]
  fx <- fixed_cancer_z[i, "mean"]
  coef <- fixed_cancer_z[i, 1]
  prop_fx <- exp(int + fx) / (1 + exp(int + fx))
  ci_025 <- exp(fixed_cancer_z[which(fixed_cancer_z$coef == "mu_z"), "X0.025quant"] + fixed_cancer_z[i, "X0.025quant"]) / 
    (1 + exp(fixed_cancer_z[which(fixed_cancer_z$coef == "mu_z"), "X0.025quant"] + fixed_cancer_z[i, "X0.025quant"]))
  ci_975 <- exp(fixed_cancer_z[which(fixed_cancer_z$coef == "mu_z"), "X0.975quant"] + fixed_cancer_z[i, "X0.975quant"]) /
    (1 + exp(fixed_cancer_z[which(fixed_cancer_z$coef == "mu_z"), "X0.975quant"] + fixed_cancer_z[i, "X0.975quant"]))
  sig <- ifelse((fixed_cancer_z[i, "X0.025quant"] > 0 & fixed_cancer_z[i, "X0.975quant"] > 0) |
                  (fixed_cancer_z[i, "X0.025quant"] < 0 & fixed_cancer_z[i, "X0.975quant"] < 0),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_cancer_z[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = coef, 
                   prob = prop_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  fx_list[[i]] <- df
}
fx_list = do.call(rbind, fx_list)
fx_list

data.frame(fixed_z[, 1], 
           prob = round(exp(fixed_z[, 2:8] ) / (1 + exp(fixed_z[, 2:8])), 3))

fixed_z

# NOT log transformed
# Gini
exp(fixed_z[1, 2:8]) / (1 + exp(fixed_z[1, 2:8]))
exp(fixed_z[3, 2:8] + fixed_z[1, 2:8]) / (1 + exp(fixed_z[3, 2:8] + fixed_z[1, 2:8]))
exp(fixed_z[3, 2:8]) / (1 + exp(fixed_z[3, 2:8]))
exp(fixed_z[4, 2:8] + fixed_z[1, 2:8]) / (1 + exp(fixed_z[4, 2:8] + fixed_z[1, 2:8]))

for(i in c(1:10)) {
  mean <- logit2prob(fixed_cancer_z[i, "mean"])
  lower <- logit2prob(fixed_cancer_z[i, "X0.025quant"])
  upper <- logit2prob(fixed_cancer_z[i, "X0.975quant"])
  df <- data.frame(i = fixed_cancer_z[i, 1],
                   mean = mean,
                   lower = lower,
                   upper = upper)
  print(df)
}

fixed_cancer_z[1, "mean"]


```

## probability for increase in simple bin interactions

```{r}
intx_list = vector("list", length = 17)
for(i in c(11:17)) {
  int <- fixed_cancer_z[which(fixed_cancer_z$coef == "mu_z"), "mean"]
  intx_name_1 <- substr(fixed_cancer_z[i, "coef"], 1, 7)
  intx_name_2 <- substr(fixed_cancer_z[i, "coef"], 9, 15)
  #print(intx_name_2)
  intx <- fixed_cancer_z[i, "mean"]
  intx_name <- ifelse(intx_name_1 == "x_pov_z", intx_name_2, intx_name_1)
  fx_1 <- fixed_cancer_z[which(fixed_cancer_z$coef == intx_name), "mean"]
  fx_2 <- fixed_cancer_z[which(fixed_cancer_z$coef == "x_pov_z"), "mean"]
  
  prop_fx <- exp(int + intx + fx_1 + fx_2) / (1 + exp(int + intx + fx_1 + fx_2))
  ci_025 <- exp(fixed_cancer_z[which(fixed_cancer_z$coef == "mu_z"), "X0.025quant"] +
                  fixed_cancer_z[i, "X0.025quant"] +
                  fixed_cancer_z[which(fixed_cancer_z$coef == intx_name), "X0.025quant"] +
                  fixed_cancer_z[which(fixed_cancer_z$coef == "x_pov_z"), "X0.025quant"]) /
    (1 + exp(fixed_cancer_z[which(fixed_z$coef == "mu_z"), "X0.025quant"] +
                  fixed_cancer_z[i, "X0.025quant"] +
                  fixed_cancer_z[which(fixed_cancer_z$coef == intx_name), "X0.025quant"] +
                  fixed_cancer_z[which(fixed_cancer_z$coef == "x_pov_z"), "X0.025quant"]))
  ci_975 <- exp(fixed_cancer_z[which(fixed_z$coef == "mu_z"), "X0.975quant"] +
                  fixed_cancer_z[i, "X0.975quant"] +
                  fixed_cancer_z[which(fixed_z$coef == intx_name), "X0.975quant"] +
                  fixed_cancer_z[which(fixed_z$coef == "x_pov_z"), "X0.975quant"]) /
    (1 + exp(fixed_cancer_z[which(fixed_z$coef == "mu_z"), "X0.975quant"] +
                  fixed_cancer_z[i, "X0.975quant"] +
                  fixed_cancer_z[which(fixed_z$coef == intx_name), "X0.975quant"] +
                  fixed_cancer_z[which(fixed_z$coef == "x_pov_z"), "X0.975quant"]))
  sig <- ifelse((fixed_cancer_z[i, "X0.025quant"] > 0 & fixed_cancer_z[i, "X0.975quant"] > 0) |
                  (fixed_cancer_z[i, "X0.025quant"] < 0 & fixed_cancer_z[i, "X0.975quant"] < 0),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_cancer_z[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = fixed_cancer_z[i, "coef"], 
                   prob = prop_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  intx_list[[i]] <- df
}
intx_list = do.call(rbind, intx_list)
intx_list


for(i in c(11:17)) {
  mean <- logit2prob(fixed_z[i, "mean"])
  lower <- logit2prob(fixed_z[i, "X0.025quant"])
  upper <- logit2prob(fixed_z[i, "X0.975quant"])
  df <- data.frame(i = fixed_z[i, 1],
                   mean = mean,
                   lower = lower,
                   upper = upper)
  print(df)
}



```

## combined binomial fixed effects

```{r}
fixed_bin_list <- rbind(fx_list, intx_list)
fixed_bin_list

write.csv(fixed_bin_list, "fixed_bin_list.csv")
```


# Gamma
## probability increase for simple gamma effects

```{r}
fx_list = vector("list", length = 11)

for(i in c(2:10)) {
  int <- fixed_cancer_y[which(fixed_cancer_y$coef == "mu_y"), "mean"]
  fx <- fixed_cancer_y[i, "mean"]
  coef <- fixed_cancer_y[i, 1]
  perc_fx <- log(int + fx)
  ci_025 <- log(fixed_cancer_y[i, "X0.025quant"] + fixed_cancer_y[which(fixed_cancer_y$coef == "mu_y"), "X0.025quant"])
  ci_975 <- log(fixed_cancer_y[i, "X0.975quant"] + fixed_cancer_y[which(fixed_cancer_y$coef == "mu_y"), "X0.975quant"])
  sig <- ifelse((fixed_cancer_y[i, "X0.025quant"] > 0 & fixed_cancer_y[i, "X0.975quant"] > 0) |
                  (fixed_cancer_y[i, "X0.025quant"] < 0 & fixed_cancer_y[i, "X0.975quant"] < 0),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_cancer_y[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = coef, 
                   percent = perc_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  fx_list[[i]] <- df
}
#fx_list_2 = do.call(rbind, fx_list)
#fx_list_2
fx_list = do.call(rbind, fx_list)
fx_list

exp(4.551151518)

data.frame(fixed_cancer_y[, 1], 
           round(exp(fixed_cancer_y[, 2:8]), 3))
fixed_cancer_y

```

## probability for interactions

```{r}
intx_list = vector("list", length = 17)
for(i in c(11:17)) {
  int <- fixed_cancer_y[which(fixed_cancer_y$coef == "mu_y"), "mean"]
  intx_name_1 <- substr(fixed_cancer_y[i, "coef"], 1, 7)
  intx_name_2 <- substr(fixed_cancer_y[i, "coef"], 9, 15)
  #print(intx_name_2)
  intx <- fixed_cancer_y[i, "mean"]
  intx_name <- ifelse(intx_name_1 == "x_pov_y", intx_name_2, intx_name_1)
  fx_1 <- fixed_cancer_y[which(fixed_cancer_y$coef == intx_name), "mean"]
  fx_2 <- fixed_cancer_y[which(fixed_cancer_y$coef == "x_pov_y"), "mean"]
  
  perc_fx <- log(fixed_cancer_y[i, "mean"] + int)# / + fx_1 + fx_2)# / (1 + exp(int + intx + fx_1 + fx_2))
  ci_025 <- log(fixed_cancer_y[i, "X0.025quant"] + fixed_cancer_y[which(fixed_cancer_y$coef == "mu_y"), "X0.025quant"])
  ci_975 <- log(fixed_cancer_y[i, "X0.975quant"] + fixed_cancer_y[which(fixed_cancer_y$coef == "mu_y"), "X0.975quant"])
  sig <- ifelse((fixed_cancer_y[i, "X0.025quant"] > 0 & fixed_cancer_y[i, "X0.975quant"] > 0) |
                  (fixed_cancer_y[i, "X0.025quant"] < 0 & fixed_cancer_y[i, "X0.975quant"] < 0),
                "sig",
                "not sig")
  dir <- ifelse(sig == "sig" & fixed_cancer_y[i, "mean"] > 0, "+", "-")
  df <- data.frame(coef = fixed_cancer_y[i, "coef"], 
                   percent = perc_fx,
                   ci_025 = ci_025,
                   ci_975 = ci_975,
                   sig = sig,
                   dir = dir)
  intx_list[[i]] <- df
}
intx_list = do.call(rbind, intx_list)
intx_list

round(res$summary.fixed, 3)

```

## combined gamma fixed effects

```{r}
fixed_gam_list <- rbind(fx_list, intx_list)
fixed_gam_list

write.csv(fixed_gam_list, "fixed_gam_list.csv")
```
