==> 06 manuscript figures.Rmd <==
---
title: "manuscript figs"
author: "Brenna Kelly"
date: "2023-03-05"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(sf)
library(tmap)
library(plyr)
library(tidyr)
library(dplyr)
library(ggpubr)
library(scales)
library(ggplot2)
library(stringr)
library(viridis)
library(ggExtra)
library(tableone)
library(gridExtra)
library(HatchedPolygons)
```

# Figure X

```{r}
tri <- read.csv("../stream-pollution-analysis/data/tri.csv") %>%
  mutate(water = ifelse(unit_of_measure == "Pounds", water, water/16))

tri_counts <- tri %>% 
  group_by(trifd, year) %>% 
  tally() # number of rows (i.e., chemicals) (per year, facility)
tri_sums <- aggregate(tri$water, by = list(tri$trifd, tri$year), FUN = sum)

tri_amts <- cbind(tri_sums, tri_counts)

tri_amts$x_log <- log(tri_amts$x + 0.00001)
tri_amts$n_log <- log(tri_amts$n)

tri_amts <- tri_amts %>%
  filter(!year %in% c(1987:1999)) # only past 20 years

tri_amts <- tri_amts[order(tri_amts$n_log), ] # sort by number, for stacked bar


my_breaks <- c(0, 0.006, 0.06, 0.6, 6)

tri_amts$x_scale <- tri_amts$x / 1000000

ggplot(data = tri_amts, mapping = aes(x = year, y = x_scale)) +
  geom_bar(stat = "identity", aes(fill = n)) +
  scale_fill_viridis_c(breaks = my_breaks, 
                       labels = my_breaks[1:5],
                       trans = scales::pseudo_log_trans(sigma = 0.06), 
                       option = "C",
                       name = "Number of \nChemicals") +
  xlab("Year") +
  ylab("Pounds of Pollution (M)") +
  theme(axis.title.x = element_text(colour = "black"),
        axis.title.y = element_text(colour = "black")) +
  theme_bw()
```

# Figure Y

```{r}

```

# Table 1: Descriptives

```{r}
pop_tox <- read.csv("data/pop_tox_merged.csv")
pop_tox$geoid <- str_pad(pop_tox$geoid, 5, pad = 0)



```

# Figure Z: Descriptives

```{r}
names(pop_tox)

#pop_tox_2020 <- pop_tox %>%
#  filter(year == 2020)

pop_tox_avg <- pop_tox %>%
  group_by(geoid) %>%
  summarise_if(is.numeric, .funs = mean) %>%
  mutate(rsei_score_bin_txt = ifelse(rsei_score_bin >= 0.5, ">0.5", "<0.5"))

pop_tox_avg$rsei_score_bin_rd <- round(pop_tox_avg$rsei_score_bin, digits = 0)
#options(scipen = 999)
vars <- c("total", "white", "black", "aian", "asian", "nhpi", "other", 
          "tom", "hisp", "pov", "rsei.score", "rsei.score.cancer",
          "rsei.score.noncancer")
tab1 <- CreateTableOne(vars, c("rsei_score_bin"), pop_tox_avg)
tab1 <- CreateTableOne(vars, c("rsei_score_bin"), pop_tox)

1 - ((nrow(pop_tox_avg) - sum(1869 + 696)) / nrow(pop_tox_avg))

prop.table(table(pop_tox_avg$rsei_score_bin == 1))
prop.table(table(pop_tox_avg$rsei_score_bin == 0))

prop.table(table(pop_tox_avg$rsei_score_bin != 0, pop_tox_avg$rsei_score_bin != 1))

tab1_print <- print(tab1, showAllLevels = TRUE, formatOptions = list(big.mark = ","))
write.csv(tab1_print, file = "table 1.csv")

```




# Figure A: Map

```{r}

geometry <- st_read("data/geometry.shp")
st <- st_read("/Users/brenna/Documents/School/GEOG 6960/Cont States/cont_states.shp")

pop_tox_20 <- pop_tox %>%
  filter(year == 2020)

dat <- merge(geometry, pop_tox_20, by = "geoid", all = TRUE)

dat$rsei.score_log <- log(dat$rsei.score + 0.9)

bbox_new <- st_bbox(geometry)

dat_bin <- dat %>%
  filter(rsei_score_bin == 0)

dat_bin_fill <- dat %>%
  filter(rsei_score_bin == 1)


#png(file = "../Manuscript/figs/2020 RSEI Map.png",
#    width = 1500, height = 900)
amt_geog$rsei.score_log <- log(amt_geog$avg_amount + 1)
tm_shape(amt_geog, bbox = bbox_new) +
  tm_polygons(col = "rsei.score_log", palette = "-plasma", 
              lwd = 0, style = "cont") +
  tm_layout(legend.width = 20,
            legend.stack = "horizontal") +
  tm_shape(dat_bin) +
  tm_polygons(col = "gray99", lwd = 0) +
  tm_shape(st) +
  tm_borders(col = "gray20", lwd = 0.5) 
#dev.off()


#
```
# Figure B: Time

```{r}

plot(aggregate(pop_tox$rsei_score_bin, by = list(pop_tox$year), FUN = mean))
plot(aggregate(pop_tox$rsei.score, by = list(pop_tox$year), FUN = mean))

plot(aggregate(pop_tox$rsei.score, by = list(pop_tox$year), FUN = max))

test <- pop_tox %>%
  group_by(year) %>%
  slice(which.max(rsei.score))
  #summarise(max = max(rsei.score, na.rm = TRUE))

test <- pop_tox
test$total_greater <- ifelse(test$rsei.score.cancer < test$rsei.score.noncancer,
                             "noncancer", "cancer")
test <- test %>%
  filter(!duplicated(geoid))

98495 + 1133646 + 207291 + 155461 + 4101752
table(test$total_greater)

x <- aggregate(pop_tox$total, by = list(pop_tox$year, pop_tox$rsei_score_bin), FUN = sum) %>%
  spread(Group.2, x)
x$ratio_exposed <- x$`1` / (x$`0` + x$`1`)

```

# summary by EPA region

```{r}

epa_sum <- aggregate(pop_tox$rsei.score, by = list(pop_tox$epa_region, pop_tox$year), FUN = mean)
names(epa_sum) <- c("epa_region", "year", "rsei.score")

ggplot(epa_sum, aes(x = year, y = rsei.score, colour = epa_region)) +
  geom_line()

```

##number of times polluted in study period

```{r}
yrs <- aggregate(pop_tox$rsei_score_bin, by = list(pop_tox$geoid), FUN = sum)
names(yrs) <- c("geoid", "count")
yrs_geog <- merge(geometry, yrs, by = "geoid", all = TRUE)

table(yrs_geog$count)
round(prop.table(table(yrs_geog$count)), 2)

# BIN number of rsei_bin by county
tm_shape(yrs_geog) +
  tm_polygons(col = "count", lwd = 0, style = "cont", palette = "-inferno",
              title = "Number of \nYears Polluted") +
  tm_shape(st) +
  tm_borders(col = "gray20", lwd = 0.5)


# GAM amount of rsei by county
amt_geog <- aggregate(pop_tox$rsei.score, by = list(pop_tox$geoid), FUN = mean, na.rm = TRUE)
names(amt_geog) <- c("geoid", "avg_amount")
amt_geog <- merge(geometry, amt_geog, by = "geoid", all = TRUE)
amt_geog$avg_amount_log <- log(amt_geog$avg_amount + 1)
tm_shape(amt_geog) +
  tm_polygons(col = "avg_amount_log", lwd = 0, style = "cont", palette = "Reds", #-plasma
              title = "Average \nRSEI Score") +
  tm_layout(legend.width = 20,
            legend.stack = "horizontal") +
  tm_shape(dat_bin) +
  tm_polygons(col = "gray99", lwd = 0) +
  tm_shape(st) +
  tm_borders(col = "gray20", lwd = 0.5)


# epa table
epa_count <- data.frame(aggregate(pop_tox$rsei_score_bin, 
                                  by = list(pop_tox$epa_region), FUN = mean), 
                       aggregate(pop_tox$rsei.score, by = list(pop_tox$epa_region),
                                 FUN = mean)) %>%
  select(-Group.1.1)
names(epa_count) <- c("epa region", "%", "mean")
epa_count$mean <- round(epa_count$mean, 2)
epa_count$`%` <- round(epa_count$`%`, 3)

epa_count

```

## Average over time
## % counties, % pop over time

```{r}
yrs <- data.frame(aggregate(pop_tox$rsei.score, by = list(pop_tox$year), FUN = mean),
                  aggregate(pop_tox$rsei.score, by = list(pop_tox$year), FUN = sd)) %>%
  select(-Group.1.1)
names(yrs) <- c("year", "avg_rsei", "sd_rsei")

ci_95 <- data.frame(ddply(pop_tox, "year", summarise, WQ95 = quantile(rsei.score, .95)),
                    ddply(pop_tox, "year", summarise, WQ95 = quantile(rsei.score, .05))) %>%
  select(-year.1)
names(ci_95) <- c("year", "ci_95", "ci_05")

yrs <- cbind(yrs, ci_95)
yrs <- yrs[, c(1:3, 5:6)]


ggplot(yrs, aes(x = year, y = avg_rsei, ymin = ci_05, ymax = ci_95)) +
  geom_line(lwd = 1) +
  geom_ribbon(alpha = 0.5) +
  theme_bw()

t.test(pop_tox$rsei.score, pop_tox$year)


pop_exp <- data.frame(aggregate(pop_tox$total, by = list(pop_tox$rsei_score_bin,
                                                         pop_tox$year), FUN = sum)) %>%
  group_by(Group.2) %>%
  spread(Group.1, x) %>%
  mutate(exposed_p = `1` / (`0` + `1`))
pop_exp$type = "Population"
names(pop_exp) <- c("year", "not_exp", "exp", "exposed_p", "type")

counties_exp <- data.frame(table(pop_tox$year, pop_tox$rsei_score_bin)) %>%
  group_by(Var1) %>%
  spread(Var2, Freq) %>%
  mutate(exposed_c = `1` / (`0` + `1`))
counties_exp$type = "Counties"
names(counties_exp) <- c("year", "not_exp", "exp", "exposed_p", "type")

counties_exp$year <- as.numeric(as.character(counties_exp$year))

exposed <- rbind(counties_exp, pop_exp)
exposed$exposed_p <- exposed$exposed_p * 100
#exposed$year <- as.factor(exposed$year)

ggplot(exposed, aes(x = year, y = exposed_p, fill = type, colour = type)) +
  geom_line(lwd = 1) +
  scale_color_manual(values=c("#E2765A", "#6B1B19")) + #"#64266F", "#d34841"
  geom_point(shape = 18, cex = 4, aes(x = year, y = exposed_p)) +
  ylim(0, 100) +
  xlab("Year") +
  ylab("% With Risk Potential") +
  ggtitle("Percent with Water Pollution Risk Potential in U.S., 2011-2020") +
  theme(legend.title = element_blank(),
        panel.background = element_rect(fill="gray95"))


ggplot(counties_exp, aes(x = year, y = exposed_p)) +
  geom_line(lwd = 1, color = "#6f7fee") +
  #scale_color_manual(values=c("#E2765A")) + #"#64266F", "#d34841"
  geom_point(shape = 18, cex = 4, aes(x = year, y = exposed_p), color = "#6f7fee") +
  xlab("Year") +
  ylab("% With Risk Potential") +
  ggtitle("Percent of U.S. Counties with Polluted Water, 2011-2020") +
  theme(legend.title = element_blank(),
        panel.background = element_rect(fill="gray95")) +
  theme_minimal()
pop_exp$exp_m <- pop_exp$exp / 1000000
ggplot(pop_exp, aes(x = year, y = exp_m)) +
  geom_line(lwd = 1, color = "#F21A00") +
  #scale_color_manual(values=c("#E2765A")) + #"#64266F", "#d34841"
  geom_point(shape = 18, cex = 4, aes(x = year, y = exp_m), color = "#F21A00") +
  xlab("Year") +
  ylab("Population (m)") +
  ggtitle("U.S. Population with Polluted Water, 2011-2020") +
  theme(legend.title = element_blank(),
        panel.background = element_rect(fill="gray95")) +
  theme_minimal()
Zissou <- c("#F21A00", "#E1AF00","#EBCC2A","#3B9AB2")
options(scipen = 999)

```


## heat map of time table
```{r}
library(data.table)

pop_sc <- pop_tox[, c(3, 42)] %>%
  mutate(rsei_range = case_when(rsei.score == 0 ~ "0",
                                rsei.score > 0 & rsei.score <= 1 ~ "<1",
                                rsei.score > 1 & rsei.score <= 10 ~ "<10",
                                rsei.score > 10 & rsei.score <= 100 ~ "<100",
                                rsei.score > 100 & rsei.score <= 1000 ~ "<1000",
                                rsei.score > 1000 & rsei.score <= 10000 ~ "<10000",
                                rsei.score > 10000 & rsei.score <= 100000 ~ "<100000",
                                rsei.score > 100000 & rsei.score <= 1000000 ~ "<1000000",
                                rsei.score > 1000000 ~ ">1000000"))

aggregate(pop_sc$rsei.score, by = list(pop_sc$year, pop_sc$rsei_range), FUN = mean)

table(pop_sc$year, pop_sc$rsei.score > 1000000)

#pop_sc$rsei.score <- log(pop_sc$rsei.score + 0.1)
#setDT(pop_sc)

#pop_sc[, as.list(summary(rsei.score)), by = year]

table(pop_tox$rsei.score <= 1 & pop_tox$rsei.score > 0)

tab <- data.frame(table(pop_sc$year, pop_sc$rsei_range))
names(tab) <- c("year", "rsei_range", "count")

#tab$year <- factor(tab$year, levels = c(2011, 2012, 2013, 2014, 2015
#                                        2016, 2017, 2018, 2019, 2020))
tab$rsei_range <- factor(tab$rsei_range, levels = c("0", "<1", "<10", "<100",
                                                    "<1000", "<10000", "<100000",
                                                    "<1000000", ">1000000"))

tab$count[tab$rsei_range == "0"] <- NA

tab_0 <- tab %>%
  filter(count < 2000)
#tab_not0 <- tab %>%
#  filter(count >= 2000)

ggplot(tab, aes(year, rsei_range, fill = count)) + 
  #geom_tile(size=0, stat="identity") + 
  #geom_tile(data = tab_not0, fill = "#14005e") +
  #geom_text(data = tab_not0, aes(label = count), colour = "white") +
  geom_tile(size = 0, stat="identity", aes(fill = count)) +
  geom_text(aes(label = count)) +
  scale_fill_gradientn(colors = viridis_pal(direction = -1, 
                                   option = "C")(50), #colors = c("white", "cyan", "darkcyan"),
                       na.value = "#0a125a", limits = c(0, 277)) +
  geom_tile(data = tab_not0, fill = "#0a125a") + #0a125a #darkslategrey
  geom_text(data = tab_not0, aes(label = count), colour = "white") +
  ggtitle("RSEI Scores 2011-2020") +
  xlab("Year") + 
  ylab("RSEI Score Range") +
  theme(legend.text = element_text(color="black", size=rel(1.3)),
        axis.text = element_text(color="black", size=rel(1.5))) +
  labs(fill = "Number of Counties \nin RSEI Range") +
  theme_minimal()
  

```



```{r}

```
